Split visitor out into its own file.
