+++ parser/parser.go
@@ -6,48 +6,17 @@ import (
 	"fmt"
 	"io"
 	"math/big"
-	"reflect"
 	"strings"
 
 	"github.com/alecthomas/participle/v2"
 	"github.com/alecthomas/participle/v2/lexer"
 )
 
-type Node interface {
-	children() []Node
-}
-
-// Visitor function.
-type Visitor func(node Node, next func() error) error
-
-// Visit all nodes in the AST.
-func Visit(root Node, visit Visitor) error {
-	return visit(root, func() error {
-		for _, child := range root.children() {
-			pv := reflect.ValueOf(child)
-			if pv.Kind() != reflect.Struct && pv.IsNil() {
-				continue
-			}
-			if err := Visit(child, visit); err != nil {
-				return err
-			}
-		}
-		return nil
-	})
-}
-
 type Proto struct {
 	Pos lexer.Position
 
 	Syntax  string   `("syntax" "=" @String ";")?`
 	Entries []*Entry `{ @@ { ";" } }`
-}
-
-func (p *Proto) children() (out []Node) {
-	for _, child := range p.Entries {
-		out = append(out, child)
-	}
-	return out
 }
 
 type Entry struct {
@@ -60,10 +29,6 @@ type Entry struct {
 	Enum    *Enum    `| @@`
 	Option  *Option  `| "option" @@`
 	Extend  *Extend  `| @@`
-}
-
-func (e *Entry) children() (out []Node) {
-	return []Node{e.Import, e.Message, e.Service, e.Enum, e.Option, e.Extend}
 }
 
 type Import struct {
@@ -79,14 +44,6 @@ type Option struct {
 	Name  []*OptionName `@@+`
 	Attr  *string       `[ @("."? Ident { "." Ident }) ]`
 	Value *Value        `"=" @@`
-}
-
-func (o *Option) children() (out []Node) {
-	for _, name := range o.Name {
-		out = append(out, name)
-	}
-	out = append(out, o.Value)
-	return
 }
 
 type OptionName struct {
@@ -108,21 +65,10 @@ type Value struct {
 	Array     *Array     `| @@`
 }
 
-func (v *Value) children() (out []Node) {
-	return []Node{v.ProtoText, v.Array}
-}
-
 type ProtoText struct {
 	Pos lexer.Position
 
 	Fields []*ProtoTextField `( @@ ( "," | ";" )? )*`
-}
-
-func (p *ProtoText) children() (out []Node) {
-	for _, field := range p.Fields {
-		out = append(out, field)
-	}
-	return
 }
 
 type ProtoTextField struct {
@@ -133,35 +79,16 @@ type ProtoTextField struct {
 	Value *Value `( ":"? @@ )`
 }
 
-func (p *ProtoTextField) children() (out []Node) {
-	out = append(out, p.Value)
-	return out
-}
-
 type Array struct {
 	Pos lexer.Position
 
 	Elements []*Value `"[" [ @@ { [ "," ] @@ } ] "]"`
 }
 
-func (a *Array) children() (out []Node) {
-	for _, element := range a.Elements {
-		out = append(out, element)
-	}
-	return
-}
-
 type Extensions struct {
 	Pos lexer.Position
 
 	Extensions []*Range `"extensions" @@ { "," @@ }`
-}
-
-func (e *Extensions) children() (out []Node) {
-	for _, rng := range e.Extensions {
-		out = append(out, rng)
-	}
-	return
 }
 
 type Reserved struct {
@@ -171,21 +98,10 @@ type Reserved struct {
 	FieldNames []string `| @String { "," @String }`
 }
 
-func (r *Reserved) children() (out []Node) {
-	for _, rng := range r.Ranges {
-		out = append(out, rng)
-	}
-	return
-}
-
 type Range struct {
 	Start int  `@Int`
 	End   *int `  [ "to" ( @Int`
 	Max   bool `           | @"max" ) ]`
-}
-
-func (r *Range) children() (out []Node) {
-	return nil
 }
 
 type Extend struct {
@@ -195,13 +111,6 @@ type Extend struct {
 	Fields    []*Field `"{" { @@ [ ";" ] } "}"`
 }
 
-func (e *Extend) children() (out []Node) {
-	for _, field := range e.Fields {
-		out = append(out, field)
-	}
-	return
-}
-
 type Service struct {
 	Pos lexer.Position
 
@@ -209,22 +118,11 @@ type Service struct {
 	Entry []*ServiceEntry `"{" { @@ [ ";" ] } "}"`
 }
 
-func (s *Service) children() (out []Node) {
-	for _, entry := range s.Entry {
-		out = append(out, entry)
-	}
-	return
-}
-
 type ServiceEntry struct {
 	Pos lexer.Position
 
 	Option *Option `  "option" @@`
 	Method *Method `| @@`
-}
-
-func (s *ServiceEntry) children() (out []Node) {
-	return []Node{s.Option, s.Method}
 }
 
 type Method struct {
@@ -238,24 +136,11 @@ type Method struct {
 	Options           Options `[ "{" { "option" @@ ";" } "}" ]`
 }
 
-func (m *Method) children() (out []Node) {
-	out = []Node{m.Request, m.Response}
-	out = append(out, m.Options.children()...)
-	return
-}
-
 type Enum struct {
 	Pos lexer.Position
 
 	Name   string       `"enum" @Ident`
 	Values []*EnumEntry `"{" { @@ { ";" } } "}"`
-}
-
-func (e *Enum) children() (out []Node) {
-	for _, enum := range e.Values {
-		out = append(out, enum)
-	}
-	return
 }
 
 type EnumEntry struct {
@@ -266,18 +151,7 @@ type EnumEntry struct {
 	Reserved *Reserved  `| "reserved" @@`
 }
 
-func (e *EnumEntry) children() (out []Node) {
-	return []Node{e.Value, e.Option, e.Reserved}
-}
-
 type Options []*Option
-
-func (o Options) children() (out []Node) {
-	for _, option := range o {
-		out = append(out, option)
-	}
-	return
-}
 
 type EnumValue struct {
 	Pos lexer.Position
@@ -288,22 +162,11 @@ type EnumValue struct {
 	Options Options `[ "[" @@ { "," @@ } "]" ]`
 }
 
-func (e *EnumValue) children() (out []Node) {
-	return e.Options.children()
-}
-
 type Message struct {
 	Pos lexer.Position
 
 	Name    string          `"message" @Ident`
 	Entries []*MessageEntry `"{" { @@ } "}"`
-}
-
-func (m *Message) children() (out []Node) {
-	for _, entry := range m.Entries {
-		out = append(out, entry)
-	}
-	return
 }
 
 type MessageEntry struct {
@@ -319,10 +182,6 @@ type MessageEntry struct {
 	Field      *Field      ` | @@ ) { ";" }`
 }
 
-func (m *MessageEntry) children() (out []Node) {
-	return []Node{m.Enum, m.Option, m.Message, m.Oneof, m.Extend, m.Reserved, m.Extensions, m.Field}
-}
-
 type OneOf struct {
 	Pos lexer.Position
 
@@ -330,22 +189,11 @@ type OneOf struct {
 	Entries []*OneOfEntry `"{" { @@ { ";" } } "}"`
 }
 
-func (o *OneOf) children() (out []Node) {
-	for _, entry := range o.Entries {
-		out = append(out, entry)
-	}
-	return
-}
-
 type OneOfEntry struct {
 	Pos lexer.Position
 
 	Field  *Field  `@@`
 	Option *Option `| "option" @@`
-}
-
-func (o *OneOfEntry) children() (out []Node) {
-	return []Node{o.Field, o.Option}
 }
 
 type Field struct {
@@ -359,10 +207,6 @@ type Field struct {
 	Direct *Direct `| @@ )`
 }
 
-func (f *Field) children() (out []Node) {
-	return []Node{f.Group, f.Direct}
-}
-
 type Direct struct {
 	Pos lexer.Position
 
@@ -373,25 +217,12 @@ type Direct struct {
 	Options Options `[ "[" @@ { "," @@ } "]" ]`
 }
 
-func (d *Direct) children() (out []Node) {
-	out = []Node{d.Type}
-	out = append(out, d.Options.children()...)
-	return
-}
-
 type Group struct {
 	Pos lexer.Position
 
 	Name    string          `"group" @Ident`
 	Tag     int             `"=" @Int`
 	Entries []*MessageEntry `"{" { @@ [ ";" ] } "}"`
-}
-
-func (g *Group) children() (out []Node) {
-	for _, entry := range g.Entries {
-		out = append(out, entry)
-	}
-	return
 }
 
 type Scalar int
@@ -454,19 +285,11 @@ type Type struct {
 	Reference *string  `| @("."? Ident { "." Ident })`
 }
 
-func (t *Type) children() (out []Node) {
-	return []Node{t.Map}
-}
-
 type MapType struct {
 	Pos lexer.Position
 
 	Key   *Type `"map" "<" @@`
 	Value *Type `"," @@ ">"`
-}
-
-func (m *MapType) children() (out []Node) {
-	return []Node{m.Key, m.Value}
 }
 
 // Parse protobuf.
